livelikit/agents/examples/voice_agents/requirements_streamlit.txt
# Streamlit Voice Agent Optimizer Requirements

# Core Streamlit and visualization
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0

# LiveKit and voice agent dependencies
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-cartesia>=0.4.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.5.0

# Additional dependencies for optimization features
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
wave

# Data processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Async and threading support
asyncio
threading

# File handling and utilities
python-dotenv>=1.0.0
glob2>=0.7
json5>=0.9.0

# Optional: For advanced audio analysis
pyaudio>=0.2.11  # Optional: for real-time audio
webrtcvad>=2.0.10  # Optional: for additional VAD options
